+++ Skyseed/src/main/java/dev/gemberkoekje/skyseed/worldgen/structure/DevStructureGenerator.java
@@ -36,6 +36,7 @@ public final class DevStructureGenerator {
             AnimalTemplates.generateInto(base.resolve("animal"));
             DungeonTemplates.generateInto(base.resolve("dungeon"));
             RuinedPortalTemplates.generateInto(base.resolve("ruined_portal"));
+            NetherFortressTemplates.generateInto(base.resolve("nether_fortress"));
             DesertTempleTemplates.generateInto(base.resolve("desert_temple"));
             JungleTempleTemplates.generateInto(base.resolve("jungle_temple"));
             WitchHutTemplates.generateInto(base.resolve("witch_hut"));